refactor: split generation authorization boundaries
