+++ electron/capabilityCore/mcpToolResults.ts
@@ -9,6 +9,8 @@
 import { ACTIVE_JOB_STATUSES } from '../productionRun/productionRunControl'
 import { stripInternalEnrichFields } from './mcpResultEnrich'
 
+export { buildToolErrorOutcome } from './mcpToolErrorResults'
+
 export type ResultLocale = 'zh-CN' | 'en'
 
 type Ctx = { locale: ResultLocale }
@@ -36,37 +38,6 @@ const RUN_STATUS_HINT: Record<string, { zh: string; en: string; nextZh: string; 
   needs_attention: { zh: '需要处理', en: 'needs attention', nextZh: '有任务卡住了，看错误详情选恢复动作', nextEn: 'A job is stuck; check the error details for recovery actions', action: 'recover' },
   completed: { zh: '已完成', en: 'completed', nextZh: '产物已保存到项目，可在 Nomi 里查看', nextEn: 'Artifacts are saved to the project; open them in Nomi', action: 'open_in_nomi' },
   cancelled: { zh: '已取消', en: 'cancelled', nextZh: '未提交的任务不计费', nextEn: 'Unsubmitted jobs are not charged', action: 'none' },
-}
-
-/** A6 已知错误码 → 人话原因 + 恢复动作（只登记确证的码，不编造；未知码原样透传）。 */
-const ERROR_HINT: Record<string, { zh: string; en: string; recover: Array<{ zh: string; en: string }> }> = {
-  asset_not_localized: {
-    zh: '参考素材还没落到本地，生成端拿不到它',
-    en: 'A referenced asset is not localized yet, so the generator cannot read it',
-    recover: [
-      { zh: '在 Nomi 里打开该节点让素材完成本地化后重试', en: 'Open the node in Nomi to finish localizing the asset, then retry' },
-    ],
-  },
-  renderer_or_provider_unknown: {
-    zh: '找不到能执行这次生成的渲染器或供应商配置',
-    en: 'No renderer or provider configuration can execute this generation',
-    recover: [
-      { zh: '用 nomi_list_models 核对可用模型后换一个', en: 'Check available models with nomi_list_models and switch' },
-      { zh: '在 Nomi 设置里补齐该供应商的接入', en: 'Complete the provider setup in Nomi settings' },
-    ],
-  },
-}
-
-/** User projection deliberately has only four actions; protocol codes stay in structuredContent for machines. */
-const USER_ACTION_HINT: Record<string, { action: string; zh: string; en: string }> = {
-  human_approval_required: { action: 'in_nomi', zh: '请在 Nomi 确认这次生成。', en: 'Confirm this generation in Nomi.' },
-  receipt_invalid: { action: 'in_nomi', zh: '这次确认已失效，请在 Nomi 重新确认。', en: 'This confirmation is no longer valid; confirm again in Nomi.' },
-  receipt_expired: { action: 'in_nomi', zh: '确认已过期，请在 Nomi 重新确认。', en: 'The confirmation expired; confirm again in Nomi.' },
-  lease_required: { action: 'reselect_project', zh: '请重新选择当前项目。', en: 'Select the current project again.' },
-  lease_invalid: { action: 'reselect_project', zh: '项目连接已失效，请重新选择当前项目。', en: 'The project connection expired; select the current project again.' },
-  project_scope_changed: { action: 'reselect_project', zh: '项目范围已变化，请重新选择项目。', en: 'The project scope changed; select the project again.' },
-  lease_expired: { action: 'reselect_project', zh: '项目连接已过期，请重新选择当前项目。', en: 'The project connection expired; select the current project again.' },
-  lease_revoked: { action: 'reselect_project', zh: '项目连接已撤销，请重新选择当前项目。', en: 'The project connection was revoked; select the current project again.' },
 }
 
 function str(value: unknown): string {
@@ -771,53 +742,4 @@ export function buildProgressStartMessage(
       .filter(Boolean).join(' · ')
   }
   return null
-}
-
-/** A6 · 错误 → 人话原因 + 恢复动作 + 诊断信息（未知错误不编内容，原样透传 message）。 */
-export function buildToolErrorOutcome(
-  toolName: string,
-  error: unknown,
-  locale: ResultLocale = 'zh-CN',
-): { text: string; outcome: Record<string, unknown> } {
-  const ctx: Ctx = { locale }
-  const message = error instanceof Error ? error.message : String(error)
-  const errorRecord = error && typeof error === 'object' ? error as Record<string, unknown> : {}
-  const structuredCode = typeof errorRecord.code === 'string'
-    ? errorRecord.code
-    : typeof errorRecord.errorCode === 'string' ? errorRecord.errorCode : null
-  const policyCode = new Set([
-    'legacy_path_forbidden', 'feature_disabled', 'phase_not_ready', 'not_ready',
-    'human_approval_required', 'receipt_invalid', 'receipt_expired',
-    'lease_required', 'lease_invalid', 'project_scope_changed', 'lease_expired', 'lease_revoked',
-  ])
-  const code = structuredCode && policyCode.has(structuredCode)
-    ? structuredCode
-    : Object.keys(ERROR_HINT).find((key) => message.includes(key)) || null
-  const policyDetails = structuredCode && policyCode.has(structuredCode)
-    ? {
-        ...(typeof errorRecord.nextAction === 'string' ? { nextAction: errorRecord.nextAction } : {}),
-        ...(typeof errorRecord.phase === 'string' ? { phase: errorRecord.phase } : {}),
-        ...(typeof errorRecord.capability === 'string' ? { capability: errorRecord.capability } : {}),
-      }
-    : {}
-  const hint = code ? ERROR_HINT[code] : null
-  const userAction = code ? USER_ACTION_HINT[code] : null
-  const recover = hint ? hint.recover.map((r) => L(ctx, r.zh, r.en)) : []
-  const text = [
-    `✗ ${userAction ? L(ctx, userAction.zh, userAction.en) : hint ? L(ctx, hint.zh, hint.en) : message}`,
-    userAction ? null : code ? `${L(ctx, '诊断', 'diagnostic')} ${code}` : null,
-    ...(!userAction ? recover.map((line, index) => `${index + 1}. ${line}`) : []),
-    !hint && toolName === 'nomi_generate'
-      ? L(ctx, '已完成的内容安全；确认模型服务与 API Key 后可重试。', 'Finished work is safe; verify the model service and API key, then retry.')
-      : null,
-  ].filter(Boolean).join('\n')
-  return {
-    text,
-    outcome: {
-      kind: 'error', tool: toolName, errorCode: code, message,
-      recoveryActions: userAction ? [L(ctx, userAction.zh, userAction.en)] : recover,
-      ...(userAction ? { nextActions: [userAction.action] } : {}),
-      ...policyDetails,
-    },
-  }
 }
